LayoutTests/http/tests/websocket/tests/hybi/broken-utf8-expected.txt

CONSOLE MESSAGE: line 0: Could not decode a text frame as UTF-8.
Test whether WebSocket aborts the connection when it receives a text frame containing broken UTF-8 data.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
